+++ changelog.md
@@ -1,13 +1,14 @@
 # Retribution 1.0.0
 
 ## Features/Improvements
-* **[Mission Generation]** Add option to configure the maximum frontline length in settings
+* **[Mission Generation]** Add option to configure the maximum front-line length in settings
 * **[Mission Generation]** Variable flight-size (2/3/4-ship) for BAI/ANTISHIP/DEAD/STRIKE (main) missions
 * **[Modding]** Support for F-15D 'Baz' mod version 1.0
 * **[Modding]** Support for Su-30 mod version 2.01B
 * **[UI]** Add livery selector to Air Wing Configurator's squadrons.
 * **[Performance]** Added performance option: Maximum front-line unit supply per control point.
 * **[Performance]** Added performance option: Disable convoys.
+* **[Performance]** Added performance option: Front-line troops prefer roads.
 
 
 # Liberation: